Prettier

+++ contracts/PriceOracle.sol
@@ -561,19 +561,23 @@ contract PriceOracle is Ownable, IPriceOracle {
 
         // Yearn vaults
         if (_isYearnVault(_tokenIn)) {
-            price = IYearnVault(_tokenIn).pricePerShare().preciseDiv(getPrice(IYearnVault(_tokenIn).token(), _tokenOut));
-            uint yvDecimals = ERC20(_tokenIn).decimals();
+            price = IYearnVault(_tokenIn).pricePerShare().preciseDiv(
+                getPrice(IYearnVault(_tokenIn).token(), _tokenOut)
+            );
+            uint256 yvDecimals = ERC20(_tokenIn).decimals();
             if (yvDecimals < 18) {
-              price = price * 10 ** (18 - yvDecimals);
+                price = price * 10**(18 - yvDecimals);
             }
             return price;
         }
 
         if (_isYearnVault(_tokenOut)) {
-            price = getPrice(_tokenIn, IYearnVault(_tokenOut).token()).preciseMul(IYearnVault(_tokenOut).pricePerShare());
-            uint yvDecimals = ERC20(_tokenOut).decimals();
+            price = getPrice(_tokenIn, IYearnVault(_tokenOut).token()).preciseMul(
+                IYearnVault(_tokenOut).pricePerShare()
+            );
+            uint256 yvDecimals = ERC20(_tokenOut).decimals();
             if (yvDecimals < 18) {
-              price = price * 10 ** (18 - yvDecimals);
+                price = price * 10**(18 - yvDecimals);
             }
             return price;
         }
